+++ src/libecalc/domain/infrastructure/energy_components/legacy_consumer/tabulated/__init__.py
@@ -1,0 +1,2 @@
+from .tabular_consumer_function import TabulatedConsumerFunction
+from .tabular_energy_function import ConsumerTabularEnergyFunction
