+++ tests/libecalc/core/consumers/conftest.py
@@ -16,11 +16,13 @@ from libecalc.domain.infrastructure.energy_components.generator_set import Gener
 from libecalc.domain.infrastructure.energy_components.generator_set.generator_set_component import (
     GeneratorSetEnergyComponent,
 )
-from libecalc.domain.infrastructure.energy_components.legacy_consumer.consumer_function.consumer_tabular_energy_function import (
+from libecalc.domain.infrastructure.energy_components.legacy_consumer.tabulated import (
     TabulatedConsumerFunction,
 )
 from libecalc.domain.infrastructure.path_id import PathID
-from libecalc.domain.process.core.tabulated import VariableExpression
+from libecalc.domain.infrastructure.energy_components.legacy_consumer.tabulated.tabular_energy_function import (
+    VariableExpression,
+)
 from libecalc.domain.regularity import Regularity
 from libecalc.dto.types import ConsumerUserDefinedCategoryType
 from libecalc.expression import Expression
